refactoring in small static modules
